# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Summer School Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List all student profiles (directory)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: college
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
        - name: major
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get own profile
      responses:
        "200":
          description: Own profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

    patch:
      operationId: updateMyProfile
      tags: [profiles]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/onboarding:
    post:
      operationId: completeOnboarding
      tags: [profiles]
      summary: Complete onboarding (first-time profile setup)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a student profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Posts ─────────────────────────────────────────────────────────────────
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List feed posts
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [announcement, resource, discussion, introduction, poll]
        - name: hashtag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/react:
    post:
      operationId: reactToPost
      tags: [posts]
      summary: Add or toggle a reaction on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  # ── Comments ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Profile:
      type: object
      required: [id, userId, fullName, college, year, major, bio, interestTags, joinedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        fullName:
          type: string
        pronouns:
          type: ["string", "null"]
        homeCity:
          type: ["string", "null"]
        college:
          type: string
        year:
          type: string
        major:
          type: string
        bio:
          type: string
        funFact:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        instagramHandle:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        interestTags:
          type: array
          items:
            type: string
        joinedAt:
          type: string
    OnboardingInput:
      type: object
      required: [fullName, college, year, major, bio, interestTags]
      properties:
        fullName:
          type: string
        pronouns:
          type: string
        homeCity:
          type: string
        college:
          type: string
        year:
          type: string
          enum: [freshman, sophomore, junior, senior, grad]
        major:
          type: string
        bio:
          type: string
        funFact:
          type: string
        photoUrl:
          type: string
        linkedinUrl:
          type: string
        instagramHandle:
          type: string
        websiteUrl:
          type: string
        interestTags:
          type: array
          items:
            type: string
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        pronouns:
          type: string
        homeCity:
          type: string
        college:
          type: string
        year:
          type: string
        major:
          type: string
        bio:
          type: string
        funFact:
          type: string
        photoUrl:
          type: string
        linkedinUrl:
          type: string
        instagramHandle:
          type: string
        websiteUrl:
          type: string
        interestTags:
          type: array
          items:
            type: string

    # ── Posts ───────────────────────────────────────────────────────────────
    Post:
      type: object
      required: [id, authorId, authorName, type, content, reactions, commentCount, isPinned, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorPhoto:
          type: ["string", "null"]
        authorCollege:
          type: ["string", "null"]
        type:
          type: string
          enum: [announcement, resource, discussion, introduction, poll]
        content:
          type: string
        linkUrl:
          type: ["string", "null"]
        hashtags:
          type: array
          items:
            type: string
        reactions:
          $ref: "#/components/schemas/ReactionCounts"
        myReaction:
          type: ["string", "null"]
        commentCount:
          type: integer
        isPinned:
          type: boolean
        isHidden:
          type: boolean
        createdAt:
          type: string
    PostInput:
      type: object
      required: [type, content]
      properties:
        type:
          type: string
          enum: [announcement, resource, discussion, introduction, poll]
        content:
          type: string
        linkUrl:
          type: string
        hashtags:
          type: array
          items:
            type: string
    PostUpdate:
      type: object
      properties:
        content:
          type: string
        linkUrl:
          type: string
        hashtags:
          type: array
          items:
            type: string
        isPinned:
          type: boolean
        isHidden:
          type: boolean
    ReactionCounts:
      type: object
      properties:
        like:
          type: integer
        fire:
          type: integer
        heart:
          type: integer
        insightful:
          type: integer
    ReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string
          enum: [like, fire, heart, insightful]

    # ── Comments ────────────────────────────────────────────────────────────
...