# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Summer School Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  userId: number;
  fullName: string;
  /** @nullable */
  pronouns?: string | null;
  /** @nullable */
  homeCity?: string | null;
  college: string;
  year: string;
  major: string;
  bio: string;
  /** @nullable */
  funFact?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  instagramHandle?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  interestTags: string[];
  joinedAt: string;
}